Moved GenericSelect, Option, Action and Divider in own modules
