+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Action/Action.js
