+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Action/README.md
@@ -1,0 +1,2 @@
+The action component can be used as child of a select component. For an example please refer to the
+documentation of the select or multi select.
