+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Action/action.scss
