+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Action/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Action from './Action';
+
+export default Action;
