+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Action/tests/Action.test.js
