+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Divider/Divider.js
