+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Divider/divider.scss
