+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Divider/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Divider from './Divider';
+
+export default Divider;
