+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Divider/tests/Divider.test.js
