+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/Label.js
