+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/README.md
@@ -1,0 +1,2 @@
+The generic select is a general select component which is used by the standard select as well as
+the multi select. For the average user using one of those is likely the better choice.
