+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/genericSelect.scss
