+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/index.js
@@ -1,0 +1,4 @@
+// @flow
+import GenericSelect from './GenericSelect';
+
+export default GenericSelect;
