+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/label.scss
