+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/overlayList.scss
