+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/GenericSelect/tests/Label.test.js
