+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Option/Option.js
@@ -2,10 +2,10 @@
 import React from 'react';
 import type {ElementRef} from 'react';
 import classNames from 'classnames';
-import {afterElementsRendered} from '../../services/DOM';
-import Icon from '../Icon';
-import Checkbox from '../Checkbox';
-import type {OptionSelectedVisualization} from './types';
+import {afterElementsRendered} from '../../services/DOM/index';
+import Icon from '../Icon/index';
+import Checkbox from '../Checkbox/index';
+import type {OptionSelectedVisualization} from '../GenericSelect/types';
 import optionStyles from './option.scss';
 
 type Props = {
@@ -84,7 +84,7 @@ export default class Option extends React.PureComponent<Props> {
         return (
             <Checkbox
                 onChange={this.handleButtonClick}
-                className={optionStyles.checkbox}
+                className={optionStyles.input}
                 checked={this.props.selected} />
         );
     }
