+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Option/index.js
@@ -1,0 +1,4 @@
+// @flow
+import Option from './Option';
+
+export default Option;
