+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Option/option.scss
@@ -45,7 +45,7 @@ $fontSize: 12px;
     line-height: 25px;
     padding-left: 35px;
 
-    .checkbox {
+    .input {
         position: absolute;
         left: 10px;
         top: 4px;
