+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Option/tests/Option.test.js
