+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/Select/index.js
@@ -1,8 +1,4 @@
 // @flow
-import Action from './Action';
-import Divider from './Divider';
-import Option from './Option';
-import GenericSelect from './GenericSelect';
 import Select from './Select';
 
-export {Action, Divider, Option, GenericSelect, Select};
+export default Select;
